refactor(config): extend ConfigBasedProvider instead of duplicating it
Review catch: commons already has ConfigBasedProvider, the SPI base every other
provider here extends — StartupScriptProvider (flight), QueryOptimizer,
IngestionTaskFactoryProvider. My ConfigProvider hand-rolled the identical
class-name/constructor reflection it already implements.

The duplication came from mirroring the wrong template: dazzleduck-sql-common
carries an OLDER StartupScriptProvider that predates ConfigBasedProvider and
open-codes the reflection, while the copy the compactor actually imports
(dazzleduck-sql-flight) extends it. I copied the stale one.

ConfigProvider now extends ConfigBasedProvider and delegates to its
load(config, prefix, default), and moves to io.dazzleduck.sql.commons.config
beside it — where the type it extends and its only implementation both live,
rather than being split across two modules for no reason.

Net -94 lines, and all 17 tests pass untouched, which is the evidence the
deleted reflection was carrying no behaviour of its own.
