+++ dazzleduck-sql-commons/src/main/java/io/dazzleduck/sql/commons/config/ConfigProvider.java
@@ -1,6 +1,7 @@
-package io.dazzleduck.sql.common;
+package io.dazzleduck.sql.commons.config;
 
 import com.typesafe.config.Config;
+import com.typesafe.config.ConfigFactory;
 
 /**
  * Supplies configuration overrides from somewhere other than the HOCON files — typically a table in
@@ -11,7 +12,6 @@ import com.typesafe.config.Config;
  * actually running?" without opening a container. Pointing the service at a table it already has a
  * connection to makes the values queryable, auditable, and changeable without a new image.
  *
- * <p>Load via {@link #load(Config)}, which honours a {@code config_provider} block:
  * <pre>{@code
  * dazzleduck_sql_compaction {
  *     minor_compaction_frequency = 1 minute      # fallback when the provider has no such key
@@ -27,27 +27,23 @@ import com.typesafe.config.Config;
  *
  * <h2>Ordering</h2>
  *
- * A provider that reads a database must run AFTER {@link StartupScriptProvider} — the startup
- * script is what attaches the catalog. Callers therefore execute the startup script first, then
- * {@link #overrides(Config)}, then build their typed config from the result.
+ * A provider that reads a database must run AFTER the startup script — that script is what attaches
+ * the catalog. Callers therefore execute the startup script first, then {@link #overrides(Config)},
+ * then build their typed config from the result.
  *
  * <h2>Semantics</h2>
  *
  * The returned {@link Config} is an OVERLAY, not a replacement: the caller applies it with the
  * file-based config as fallback, so a key the provider does not supply keeps its file value. That
- * is what lets a deployment adopt this incrementally, and what makes the bundled
+ * is what lets a deployment adopt this incrementally, and what keeps the bundled
  * {@code application.conf} a meaningful set of defaults rather than dead weight.
  *
- * <p>Absent by default. With no {@code config_provider} block, {@link #load} returns a provider
- * that supplies nothing, and behaviour is byte-identical to before this interface existed.
+ * <p>Absent by default: with no {@code config_provider} block {@link #load} returns {@link #NONE},
+ * and behaviour is byte-identical to before this interface existed.
  */
-public interface ConfigProvider {
+public interface ConfigProvider extends ConfigBasedProvider {
 
     String CONFIG_PROVIDER_PREFIX = "config_provider";
-
-    /** Called by {@link #load} for a no-arg implementation; the block itself, not the root config. */
-    default void setConfig(Config config) {
-    }
 
     /**
      * The overrides to overlay on the file-based config, possibly empty. Never null.
@@ -62,31 +58,26 @@ public interface ConfigProvider {
     Config overrides(Config serviceConfig) throws Exception;
 
     /** Supplies nothing — the default when no {@code config_provider} block is present. */
-    ConfigProvider NONE = serviceConfig -> com.typesafe.config.ConfigFactory.empty();
+    ConfigProvider NONE = new ConfigProvider() {
+        @Override
+        public void setConfig(Config config) {
+        }
+
+        @Override
+        public Config overrides(Config serviceConfig) {
+            return ConfigFactory.empty();
+        }
+    };
 
     /**
-     * Loads a {@link ConfigProvider} from {@code config}'s {@code config_provider} block, or
-     * {@link #NONE} when the block is absent or names no {@code class}.
+     * Loads the provider named by {@code config}'s {@code config_provider} block, or {@link #NONE}
+     * when the block is absent or names no {@code class}.
      *
-     * <p>Mirrors {@link StartupScriptProvider#load}: a {@code Config}-taking constructor is
-     * preferred, falling back to a no-arg constructor plus {@link #setConfig}.
+     * <p>Delegates to {@link ConfigBasedProvider#load(Config, String, ConfigBasedProvider)} — the
+     * shared class-name/constructor resolution every other provider SPI here uses — rather than
+     * repeating that reflection.
      */
     static ConfigProvider load(Config config) throws Exception {
-        if (!config.hasPath(CONFIG_PROVIDER_PREFIX)) {
-            return NONE;
-        }
-        Config providerConfig = config.getConfig(CONFIG_PROVIDER_PREFIX);
-        if (!providerConfig.hasPath("class")) {
-            return NONE;
-        }
-        String className = providerConfig.getString("class");
-        try {
-            var ctor = Class.forName(className).getConstructor(Config.class);
-            return (ConfigProvider) ctor.newInstance(providerConfig);
-        } catch (NoSuchMethodException e) {
-            var provider = (ConfigProvider) Class.forName(className).getConstructor().newInstance();
-            provider.setConfig(providerConfig);
-            return provider;
-        }
+        return ConfigBasedProvider.load(config, CONFIG_PROVIDER_PREFIX, NONE);
     }
 }
