+++ dazzleduck-sql-commons/src/test/java/io/dazzleduck/sql/commons/TableConfigProviderTest.java
@@ -2,7 +2,7 @@ package io.dazzleduck.sql.commons;
 
 import com.typesafe.config.Config;
 import com.typesafe.config.ConfigFactory;
-import io.dazzleduck.sql.common.ConfigProvider;
+import io.dazzleduck.sql.commons.config.ConfigProvider;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
